pdf_simplifier/requirements.txt
pypdf2~=3.0.1
pdfquery~=0.4.3
pandas~=2.1.1
tabula-py~=2.8.2